+++ CHANGELOG.md
@@ -10,11 +10,13 @@ Every change is marked with Pull Request ID.
 - Added "default" option for extensions, similar to list content #328
 - Added info message if there are unpublished statusreports #340
 - Added published/unpublished indicators for statusreports in dropdown and ribbon #341
+- Added possiblity to delete unpublished statusreports #343
 
 ## Fixed
 
 - Improved failure handling for PlannerConfiguration task in Project Setup #329
 - Support adding AD groups to get porfolio insights from SP group #332
+- Change to latest statusreport when creating a new statusreport #343
 
 ## 1.2.3 - 2020-10-07